Project P6 step 3 and 4 Delete sauce controllers

+++ controllers/sauce.js
@@ -1,4 +1,5 @@
 const Sauce = require('../models/Sauce');
+const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
    const sauceObject = JSON.parse(req.body.sauce);
@@ -41,9 +42,22 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json({message: 'Objet supprimé'}))
-    .catch(error => res.status(400).json({error}));
+   Sauce.findOne({ _id: req.params.id})
+       .then(sauce => {
+           if (sauce.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = sauce.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Sauce.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 exports.getOneSauce = (req, res, next) => {
